+++ CentralStation/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Fixed destinations from old Bus Locations and Train Station content packs not always hidden when they should be (e.g. when already in their location).
+
 ## 1.3.0
 Released 06 June 2025 for SMAPI 4.1.10 or later.